+++ CHANGELOG.md
@@ -558,6 +558,8 @@
 - [Convert large longs to doubles, safely, for host calls][4099]
 - [Consistent ordering with comparators](4067)
 - [Profile engine startup][4110]
+- [Report type of polyglot values][4111]
+- [Engine can now recover from serialization failures][5591]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -650,6 +652,8 @@
 [4099]: https://github.com/enso-org/enso/pull/4099
 [4067]: https://github.com/enso-org/enso/pull/4067
 [4110]: https://github.com/enso-org/enso/pull/4110
+[4111]: https://github.com/enso-org/enso/pull/4111
+[5591]: https://github.com/enso-org/enso/pull/5591
 
 # Enso 2.0.0-alpha.18 (2021-10-12)